リファクタ: データベース接続コードを整理・シンプル化
